+++ apps/c/volna/NumericalFluxes_1.h
@@ -1,48 +1,48 @@
-void NumericalFluxes_1(double *LeftFacetValues, //OP_READ
-           double *RightFacetValues, //OP_READ
-           double *out, //OP_WRITE
-           double *FacetVolumes, //OP_READ,
-           double *Normals, //OP_READ,
-           double *maxFacetEigenvalues //OP_WRITE
+void NumericalFluxes_1(float *LeftFacetValues, //OP_READ
+           float *RightFacetValues, //OP_READ
+           float *out, //OP_WRITE
+           float *FacetVolumes, //OP_READ,
+           float *Normals, //OP_READ,
+           float *maxFacetEigenvalues //OP_WRITE
 )
 {
-  double cL = sqrt(params_g * LeftFacetValues[0]);
+  float cL = sqrt(params_g * LeftFacetValues[0]);
   cL = cL > 0.0 ? cL : 0.0;
-  double cR = sqrt(params_g * RightFacetValues[0]);
+  float cR = sqrt(params_g * RightFacetValues[0]);
   cR = cR > 0.0 ? cR : 0.0;
 
-  double uLn = LeftFacetValues[1] * Normals[0] + LeftFacetValues[2] * Normals[1];
-  double uRn = RightFacetValues[1] * Normals[0] + RightFacetValues[2] * Normals[1];
+  float uLn = LeftFacetValues[1] * Normals[0] + LeftFacetValues[2] * Normals[1];
+  float uRn = RightFacetValues[1] * Normals[0] + RightFacetValues[2] * Normals[1];
 
-  double unStar = 0.5 * (uLn + uRn) - 0.25* (cL+cR);
-  double cStar = 0.5 * (cL + cR) - 0.25* (uLn+uRn);
+  float unStar = 0.5 * (uLn + uRn) - 0.25* (cL+cR);
+  float cStar = 0.5 * (cL + cR) - 0.25* (uLn+uRn);
 
-  double sL = (uLn - cL) < (unStar - cStar) ? (uLn - cL) : (unStar - cStar);
-  double sLMinus = sL < 0.0 ? sL : 0.0;
+  float sL = (uLn - cL) < (unStar - cStar) ? (uLn - cL) : (unStar - cStar);
+  float sLMinus = sL < 0.0 ? sL : 0.0;
 
-  double sR = (uRn + cR) > (unStar + cStar) ? (uRn + cR) : (unStar + cStar);
-  double sRPlus = sR > 0.0 ? sR : 0.0;
+  float sR = (uRn + cR) > (unStar + cStar) ? (uRn + cR) : (unStar + cStar);
+  float sRPlus = sR > 0.0 ? sR : 0.0;
 
-  sL = LeftFacetValues[0] < EPS ? uRn - 2.0*cR : sL; // is this 2.0 or 2? (i.e. double/int)
+  sL = LeftFacetValues[0] < EPS ? uRn - 2.0*cR : sL; // is this 2.0 or 2? (i.e. float/int)
   sR = LeftFacetValues[0] < EPS ? uRn + cR : sR;
 
-  sR = RightFacetValues[0] < EPS ? uLn + 2.0*cL : sR; // is this 2.0 or 2? (i.e. double/int)
+  sR = RightFacetValues[0] < EPS ? uLn + 2.0*cL : sR; // is this 2.0 or 2? (i.e. float/int)
   sL = RightFacetValues[0] < EPS ? uLn - cL : sL;
 
-  double sRMinussL = sRPlus - sLMinus;
+  float sRMinussL = sRPlus - sLMinus;
   sRMinussL = sRMinussL < EPS ? EPS : sRMinussL;
 
-  double t1 = sRPlus / sRMinussL;
+  float t1 = sRPlus / sRMinussL;
   //assert( ( 0 <= t1 ) && ( t1 <= 1 ) );
 
-  double t2 = ( -1.0 * sLMinus ) / sRMinussL;
+  float t2 = ( -1.0 * sLMinus ) / sRMinussL;
   //assert( ( 0 <= t2 ) && ( t2 <= 1 ) );
 
-  double t3 = ( sRPlus * sLMinus ) / sRMinussL;
+  float t3 = ( sRPlus * sLMinus ) / sRMinussL;
 
-  double LeftFluxes_H, LeftFluxes_U, LeftFluxes_V;
+  float LeftFluxes_H, LeftFluxes_U, LeftFluxes_V;
   //inlined ProjectedPhysicalFluxes(LeftFacetValues, Normals, params, LeftFluxes);
-  double HuDotN = (LeftFacetValues[0] * LeftFacetValues[1] * Normals[0]) +
+  float HuDotN = (LeftFacetValues[0] * LeftFacetValues[1] * Normals[0]) +
           (LeftFacetValues[0] * LeftFacetValues[2] * Normals[1]);
 
   LeftFluxes_H = HuDotN;
@@ -53,9 +53,9 @@ void NumericalFluxes_1(double *LeftFacetValues, //OP_READ
   LeftFluxes_V += (.5 * params_g * Normals[1] ) * ( LeftFacetValues[0] * LeftFacetValues[0] );
   //end of inlined
 
-  double RightFluxes_H, RightFluxes_U, RightFluxes_V;
+  float RightFluxes_H, RightFluxes_U, RightFluxes_V;
   //inlined ProjectedPhysicalFluxes(RightFacetValues, Normals, params, RightFluxes);
-  double HuDotN = (RightFacetValues[0] * RightFacetValues[1] * Normals[0]) +
+  float HuDotN = (RightFacetValues[0] * RightFacetValues[1] * Normals[0]) +
           (RightFacetValues[0] * RightFacetValues[2] * Normals[1]);
 
   RightFluxes_H =   HuDotN;
@@ -89,7 +89,7 @@ void NumericalFluxes_1(double *LeftFacetValues, //OP_READ
   out[2] *= *FacetVolumes;
   out[3] = 0.0;
 
-  double maximum = abs(uLn + cL);
+  float maximum = abs(uLn + cL);
   maximum = maximum > abs(uLn - cL) ? maximum : abs(uLn - cL);
   maximum = maximum > abs(uRn + cR) ? maximum : abs(uRn + cR);
   maximum = maximum > abs(uRn - cR) ? maximum : abs(uRn - cR);
